Add reverseNegativeNumber and isPrime methods
Add two new methods to the week3 package: one for reversing negative numbers, and one to check if a number is prime. The reverseNegativeNumber method takes an integer as input, reverses the digits and leave the negative sign in front. The isPrime method checks if an input number is prime.