Add solutions for permutation combinations and unique integers summing to zero
Added core methods to generate all possible permutations of an array and to calculate N unique integers that add up to zero. For permutation, a recursive approach is used where characters are swapped and backtracked. For unique integers, positive and negative pairs are used to sum up to zero. In case of odd N, zero is added to balance the sum.